[ICU-13725] Additional MSVC compatibility fixes. X-Commit-URL: https://ssl.icu-project.org/trac/changeset/41272 Former-commit-id: 3ed43546987812518ed94a7bb3eed324268b4beb

+++ icu4c/source/i18n/number_decimalquantity.cpp
@@ -509,9 +509,9 @@ int64_t DecimalQuantity::toLong(bool truncateIfOverflow) const {
         result = result * 10 + getDigitPos(magnitude - scale);
     }
     if (isNegative()) {
-        return -result;
+        return static_cast<int64_t>(-result);
     }
-    return result;
+    return static_cast<int64_t>(result);
 }
 
 uint64_t DecimalQuantity::toFractionLong(bool includeTrailingZeros) const {
